# improve getHashCode by removing fmt.Sprintf