I'm using this library do perform quite a bit of ancestor query but is hitting high cpu usage.
did profiling and found out that getHashCode is rather expensive.

<img width="1915" alt="image" src="https://github.com/user-attachments/assets/b9badd67-6f1a-4d1e-9e65-5c5a6ceb6a57" />

did some further research and found out that sprintf can be expensive as shown in the profiling results
https://medium.com/swlh/bad-go-frivolous-sprintf-2ad28fedf1a0
https://hackernoon.com/fmtsprintf-looks-simple-but-will-burn-a-hole-in-your-pocket

# Before

##  Benchmark

```
goos: darwin
goarch: arm64
pkg: github.com/antchfx/xpath
cpu: Apple M3 Pro
=== RUN   Benchmark_GetHashCode
Benchmark_GetHashCode
=== RUN   Benchmark_GetHashCode/getHashCode
Benchmark_GetHashCode/getHashCode
Benchmark_GetHashCode/getHashCode-12             2976260               402.4 ns/op           144 B/op         10 allocs/op
PASS
ok      github.com/antchfx/xpath        2.555s
```

# After

##  Benchmark

```
goos: darwin
goarch: arm64
pkg: github.com/antchfx/xpath
cpu: Apple M3 Pro
=== RUN   Benchmark_GetHashCode
Benchmark_GetHashCode
=== RUN   Benchmark_GetHashCode/getHashCode
Benchmark_GetHashCode/getHashCode
Benchmark_GetHashCode/getHashCode-12             9322789               122.1 ns/op            88 B/op          2 allocs/op
PASS
ok      github.com/antchfx/xpath        2.000s
```